renamed folder containing loader service
