+++ client/src/app/core/loading-interceptor/loading.interceptor.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { HttpRequest, HttpHandler, HttpEvent, HttpInterceptor } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { finalize } from 'rxjs/operators';
-import { LoaderService } from '../../shared/load-controller/loader.service';
+import { LoaderService } from '../../shared/loader-service/loader.service';
 
 @Injectable()
 export class LoadingInterceptor implements HttpInterceptor {
