+++ client/src/app/shared/spinner/spinner.component.spec.ts
@@ -1,7 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { SpinnerComponent } from './spinner.component';
-import { LoaderService } from 'app/shared/load-controller/loader.service';
+import { LoaderService } from '../loader-service/loader.service';
 
 describe('SpinnerComponent', () => {
   let component: SpinnerComponent;
